+++ milkman-jdbc/src/main/java/milkman/plugin/jdbc/AbstractJdbcProcessor.java
@@ -1,0 +1,30 @@
+package milkman.plugin.jdbc;
+
+import java.sql.ResultSet;
+import java.sql.ResultSetMetaData;
+import java.sql.SQLException;
+import java.util.LinkedList;
+import java.util.List;
+
+import milkman.plugin.jdbc.domain.RowSetResponseAspect;
+
+public class AbstractJdbcProcessor {
+
+	protected void extractRows(ResultSet resultSet, RowSetResponseAspect rowSetAspect) throws SQLException {
+		ResultSetMetaData metaData = resultSet.getMetaData();		
+		
+		List<String> columnNames = new LinkedList<String>();
+		for(int i = 1; i <= metaData.getColumnCount(); ++i) { // column idx starts at 1
+			columnNames.add(metaData.getColumnName(i));
+		}
+		rowSetAspect.setColumnNames(columnNames);
+		
+		while(resultSet.next()) {
+			List<Object> row = new LinkedList<Object>();
+			for(int i = 1; i <= metaData.getColumnCount(); ++i) {// column idx starts at 1
+				row.add(resultSet.getObject(i));
+			}
+			rowSetAspect.addRow(row);
+		}
+	}
+}