+++ milkman/src/main/java/milkman/ctrl/RequestExecutor.java
@@ -1,5 +1,7 @@
 package milkman.ctrl;
 
+import java.util.Optional;
+
 import org.apache.commons.lang3.exception.ExceptionUtils;
 
 import javafx.concurrent.Service;
@@ -8,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import milkman.domain.RequestContainer;
 import milkman.domain.ResponseContainer;
+import milkman.ui.plugin.CustomCommand;
 import milkman.ui.plugin.RequestTypePlugin;
 import milkman.ui.plugin.Templater;
 
@@ -18,6 +21,7 @@ public class RequestExecutor extends Service<ResponseContainer> {
 	private final RequestContainer request; 
 	private final RequestTypePlugin plugin;
 	private final Templater templater;
+	private final Optional<CustomCommand> customCommand;
 	
 	@Override
 	protected Task<ResponseContainer> createTask() {
@@ -26,9 +30,14 @@ public class RequestExecutor extends Service<ResponseContainer> {
 			@Override
 			protected ResponseContainer call() {
 				try {
-					log.info("Execution request");
-					ResponseContainer result = plugin.executeRequest(request, templater);
-					return result;
+					if (customCommand.isPresent()) {
+						String commandId = customCommand.get().getCommandId();
+						log.info("Execute custom command: " + commandId);
+						return plugin.executeCustomCommand(commandId, request, templater);
+					} else {
+						log.info("Execute request");
+						return plugin.executeRequest(request, templater);
+					}
 				} catch (Throwable e) {
 					log.error("Execution of request failed", e);
 					String rootCauseMessage = ExceptionUtils.getRootCauseMessage(e);