+++ milkman/src/main/java/milkman/ui/plugin/CustomCommand.java
@@ -1,0 +1,9 @@
+package milkman.ui.plugin;
+
+import lombok.Value;
+
+@Value
+public class CustomCommand {
+	String commandId;
+	String commandText;
+}