+++ changelog.md
@@ -81,7 +81,9 @@ version 2.2.0
   * generic postman import for dumps, collections (v1, v2.1), environments
   * postman export
   * enabled ssh credentials for git sync
-  * jsch replaced with sshd for jgit (to support new ed25519 keys)
+    * jsch replaced with sshd for jgit (to support new ed25519 keys)
+  * added custom commands for plugins
+    * SHOW TABLES command for jdbc plugin
  
    
 Planned: